Related work items: #107699, #116999, #117741, #119681, #123560, #123619, #123993, #124374

+++ src/main/java/com/dpw/runner/shipment/services/dto/request/ShipmentRequest.java
@@ -43,9 +43,9 @@ public class ShipmentRequest extends CommonRequest implements IRunnerRequest {
     private List<FileRepoRequest> fileRepoList;
     private String financeClosedBy;
     private LocalDateTime financeClosedOn;
-    private Integer freightLocal;
+    private BigDecimal freightLocal;
     private String freightLocalCurrency;
-    private Integer freightOverseas;
+    private BigDecimal freightOverseas;
     private String freightOverseasCurrency;
     private String goodsDescription;
     private String houseBill;
